% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{stat500_ci_mean}
\alias{stat500_ci_mean}
\title{Get confidence interval of a mean}
\usage{
stat500_ci_mean(x, s, n, level = 0.95)
}
\arguments{
\item{x}{\code{numeric} Sample mean}

\item{s}{\code{numeric} Sample standard deviation}

\item{n}{\code{integer} Sample size - will be coerced to an integer}

\item{level}{\code{numeric} Confidence level - must be between 0 and 1}
}
\value{
a \code{numeric} vector
}
\description{
Calculates a confidence interval for a mean, given a confidence level between 0 and 1, a sample size, a sample mean, and a sample standard
error (estimate of population standard deviation). Takes one value for each argument and returns a vector containing the margin of error and
confidence interval. This function assumes the sampling distribution is normal.
}
\examples{
\dontrun{
stat500_ci_mean(x = 24, s = 1.7, n = 41)
stat500_ci_mean(x = 11, s = 19, n = 36, level = 0.99)
stat500_ci_mean(x = 273, s = 56.2, n = 25, level = 0.99)
}
}
