% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{stat500_ci_prop}
\alias{stat500_ci_prop}
\title{Get confidence interval of a proportion}
\usage{
stat500_ci_prop(p, n, level = 0.95, approximate_cutoff = 5)
}
\arguments{
\item{p}{\code{numeric} Sample proportion}

\item{n}{\code{integer} Sample size - will be coerced to an integer}

\item{level}{\code{numeric} Confidence level - must be between 0 and 1}

\item{approximate_cutoff}{\code{numeric} Cutoff for using the Z-method when \code{p >= approximate_cutoff && (1 - p) >= approximate_cutoff}}
}
\value{

}
\description{
Get confidence interval of a proportion
}
\examples{
\dontrun{
stat500_ci_prop(p = 0.55, n = 600, level = 0.98)
}
}
