% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{stat500_n_mean}
\alias{stat500_n_mean}
\title{Get the sample size required for a desired margin of error of a mean}
\usage{
stat500_n_mean(E, sigma, level = 0.95)
}
\arguments{
\item{E}{\code{numeric} Desired effect size/margin of error}

\item{sigma}{Standard error of the sample mean}

\item{level}{Confidence level}
}
\value{
\code{integer}
}
\description{
Get the sample size required for a desired margin of error of a mean
}
\examples{
\dontrun{
stat500_n_mean(E = 0.5, sigma = 2)
}
}
