% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{stat500_n_prop}
\alias{stat500_n_prop}
\title{Get the sample size required for a desired margin of error of a proportion}
\usage{
stat500_n_prop(E, p = 0.5, level = 0.95)
}
\arguments{
\item{E}{\code{numeric} Desired effect size/margin of error}

\item{p}{\code{numeric} Sample proportion}

\item{level}{\code{numeric} Confidence level}
}
\value{
\code{integer}
}
\description{
Get the sample size required for a desired margin of error of a proportion
}
\examples{
\dontrun{
stat500_n_prop(E = 0.01, p = 0.72)
stat500_n_prop(E = 0.01)
}
}
